refactor: move fragment validation beside the fragment checker
validateFragmentsAgainstIds compares requested fragments against the ids a
page offers, which is fragment semantics rather than parsing, so it sat
oddly in links.ts among getLinks, getCssLinks and extractFragmentIds. It
moves to fragments.ts, next to its only caller. extractFragmentIds and
isSoftNotFound stay behind, since both answer questions about HTML bytes
like the rest of that module.

validateFragments, the parse-and-compare wrapper added in 1ac4526, is
removed: no caller has been left in src since FragmentChecker took over, and
moving the comparison out from under it would have made links.ts import from
fragments.ts, which already imports links.ts. Its two tests are rewritten
against extractFragmentIds plus validateFragmentsAgainstIds - the same pair
FragmentChecker.validate uses - so both intents survive, now exercising the
production path instead of a wrapper nothing called.
